requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.1.4
django-environ==0.12.0
django-modeltranslation==0.19.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
